Integrate Arkham to mainline
BZ: 127460

Arkham is a deeply integrated security container capability for
Android for the purpose of isolation and management in mixed used
cases. Primarily Arkham is indented to enable separation of work and
personal content and capability.

Detailed information about the Arkham requirements and architecture can
be found here: https://sharepoint.amr.ith.intel.com/sites/SDI/Arkham/

This patch contains the bulk of the implementation for the Arkham
public components:

() frameworks - contains the implementation for the hooks added in the
frameworks/base project. It has two directories: enabled which has the
full implementation and disabled which has empty implementation. This
allows us to disable or enable the Arkham feature at build time.

() config - temporarily code to export the state of the Arkham feature
(enabled or disabled) to the java code. Should be moved in an Arkham
independent location.

() system/vold - contains the implementation for the ecryptfs based
encryption commands. The following commands are supported:

ess create - creates a new key suitable for ecryptfs encryption; the
key is encrypted with a hash of the user's password

ess mount <dir> - mounts ecryptfs for the given directory on top of
itself. Applications who have access to the given director will be
able to read the data in clear text form but that information is
volatile. The actual data is stored on the backing store filesystem in
encrypted form.

ess umount <dir> - unmounts the given directory

ess unmountall - unmounts all encrypts directories

ess changepw - decrypts the ecryptfs key with the old user password
and re-encrypts it with the new user password

ess delete - deletes the keys the kernel ring; called when the
container is wiped

ess disable - disabled the container key and thus breaks all container
specific encryption until the key is re-enabled; only available for
Chaabi implementation

ess enable - the reverse of the previous command; only available for
Chaabi implementation

() tools/cm - a debug tool to list containers, enable, disable
containers, install applications to container, change whitelist, etc.

Category: feature differentiation
Domain: AOSP.Extend-Arkham
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: Idd15e26000d3e2f59bb7dfde53e61b81958d43ff
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-on: http://android.intel.com:8080/126915
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>