Arkham bugfixes WW47
BZ: 156819

* Delete Container APKs while wiping container.
apk is not deleted after wiping container if it is installed only inside the container.

* Protect broadcasts among various Arkham components
Add a permission to protect Arkham specific broadcasts.
These broadcasts are used internally for various features.
In order to prevent a rogue app from sending these broadcasts,
a new permission is defined.

* Disable unsupported container lock options
Fixed a bug where, after rebooting the phone, the user would be
able to pick any lock option for the container, which would
result in container corruption.
This was caused by CMS calling DPM's setActiveAdmin method, which
would overwrite the device_policies.xml file with nothing but
the device admin's name. The setActiveAdmin call has been moved to
the extended DPM and is called after a succesful password input
has been recorded (the container has been unlocked).

* Fix contact fields whitelist for 'Normal' mode
Fixed a bug where the contact fields whitelist wasn't working for
the 'Normal' contacts merge option. This was caused by an improper
instantiation of the ContainerPolicyManager.
The Contacts Provider will now filter the fetched remote contacts
using the MDM's contacts field whitelist policy, when using the
'Normal' merge mode.

* Show progress bar in container launcher till apps are setup
Allowing the launcher to be launched before apps setup results in
showing all apps for a small duration of time.
To avoid this, show progress screen until apps are setup.
Also, non-UI apps setup is a part of this flow.

* Enforce container creation limit
Restrict the allowed number of created containers to the
limit set by the corresponding system property.
For Arkham 1.0, this limit will be set to 1.
This patch adds String constants that define the system
property names:
- Arkham enabled: ro.intel.arkham.enabled
- Maximum number of containers: ro.intel.arkham.maxcontainers

* Klocwork fixes (sprint 25)
Fixes for the Klocwork critical issues reported for the
sprint 25 daily code scan.

* Fix password expiration notification bug
Fixed a bug that was preventing password expiration notifications to be
shown. This bug was caused by the DPM failing to reload container device
admin information after the container was unlocked.
The activateContainerDeviceAdmin was removed, because the DPM's setActiveAdmin
method that it was calling was writing "empty" device admin info, leaving out
details such as password expiration time. Removed the reportSuccessfulPasswordAttempt
override from the EDPM, as it is no longer needed.
Overriden the DPM's handlePasswordExpirationNotification method to force
the DPM to reload container device admin info from xml after the BOOT_COMPLETED
intent is sent by the CMS.

* Add CM API to get a list of installed apps
Add an API to container manager to fetch the list of installed apps inside a container.

* ContentService: Restrict interaction between container and container owner
Allow only com.android.providers.userdictionary application and
any other application started by system or root to interact between container
and container owner.

* VOLD Chaabi refactorings
The container key signature is changed to use a hash of
(cid|salt) instead of hash of (key) so that we can easily find the key
associated with a container.
Vold change password flow is changed so that it does
not require the old password. The vold "ess changepw" receives only
container id, new password and salt. Based on cid and salt the container
key signature Sig = hash(cid|salt) is computed and the key is retrieved
from the kernel ring, then is encrypted with the new password and salt
and returned.

* VOLD: link with libcc54 and issue Chaabi calls
* Disable keyguard if FLAG_KEEP_SCREEN_ON is set
FLAG_KEEP_SCREEN_ON is held by activities which don't want to turn the screen off.
Examples are Youtube, Video player etc.
Disable keyguard if this flag is set so that videos can be played without keyguard interruption.

* Allow Bluetooth management and sharing in container
Redirect intents received from Bluetooth app in container.

* Fix deadlock when changing IME
Synchronize equeNotificationInternal by AMS instance.

* Remove reboot device from container public API
Remove rebootDevice method from interface.

* ContainerConstants documentation review
Moved the MergeContacts enum from ContainerCommons to ContainerConstants.
Hidden the constants that are used internally by the ContainerManagerService,
internal MDM, or extended classes.
Added description for the public constants that were lacking it.
Hidden the ContainerCommons class.
Added description to ContainerInfo methods and hid some of its methods that should not be exposed.

* VOLD,ESS: HWK error codes provided to CMH
Ess enable/disable methods updated to deliver the
HWK error codes to the ContainerMountHelper in order
to refine the errors handling on ContainerMountHelper
side.

* Do not send broadcasts throughout the system
Broadcasts should be only be sent to packages that need them.
Changing broadcasts that are sent throughout the system.

* Revert "Review enforcePermission in ContainerManagerService"
Now the PERMISSION_ADMIN is more restrictive than PERMISSION_LAUNCHER.

* Fix unprotected APIs in Container Manager
getContainerAccounts is removed from ContainerManager
and from IContainerManager.aidl. It is needed by
ExtendAccountsManagerService only which doesn't need to
go through Container Manager. So public interface is removed.

* Protect getContainerAccounts and avoid making it public API
Use reflection to call getContainerAccounts()

* Enforce complex password policies
- Added description for password policies that weren't formerly used;
- Removed some policies that aren't used;
- Added a bit of description for some policies usable by
  ContainerManager.setContainerPolicy

Change-Id: I34c82e0dac2cfa0250ebedd0831f89f4b88072e0
Category: feature differentiation
Domain: AOSP.Security-Arkham
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Mihai Serban <mihai.serban@intel.com>